+++ installer.py
@@ -385,12 +385,13 @@ def git(arg: str, folder: str | None= None, ignore: bool = False, optional: bool
 
 # reattach as needed as head can get detached
 def branch(folder=None):
-    # if args.experimental:
-    #    return None
+    if args.experimental or args.skip_git or args.skip_all:
+        return None
     t_start = time.time()
     if not os.path.exists(os.path.join(folder or os.curdir, '.git')):
         return None
     branches = []
+    detached = False
     try:
         b = git('branch --show-current', folder, optional=True)
         if b == '':
@@ -399,20 +400,27 @@ def branch(folder=None):
         if len(branches) > 0 and len(marked) > 0:
             b = marked[0]
             if ('detached' in b or 'HEAD' in b) and len(branches) > 1:
+                detached = True
                 b = branches[1].strip()
-                log.debug(f'Git detached head detected: folder="{folder}" reattach={b}')
+                log.debug(f'Submodule: folder="{folder}" reattach={b} git detached head detected')
     except Exception:
         b = git('git rev-parse --abbrev-ref HEAD', folder, optional=True)
+
     if 'main' in b:
-        b = 'main'
+        tgt = 'main'
     elif 'master' in b:
-        b = 'master'
+        tgt = 'master'
     else:
-        b = b.split('\n')[0].replace('*', '').strip()
-    log.debug(f'Submodule: folder="{folder}" branch="{b}"')
-    git(f'checkout {b}', folder, ignore=True, optional=True)
+        tgt = b.split('\n')[0].replace('*', '').strip()
+    if (tgt != b) or detached:
+        log.debug(f'Submodule: folder="{folder}" branch="{b}" target="{tgt}"')
+        git(f'checkout {tgt}', folder, ignore=True, optional=True)
+        git('fetch', folder, ignore=True)
+        git(f'merge --ff-only origin/{tgt}', folder, ignore=True)
+    else:
+        log.debug(f'Submodule: folder="{folder}" branch="{b}"')
     ts('branch', t_start)
-    return b
+    return tgt
 
 
 # restart process