+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Change Log for SD.Next
 
-## Update for 2026-09-14
+## Update for 2026-09-15
 
-### Highlights for 2026-09-14
+### Highlights for 2026-09-15
 
 *What's New*? Well, code-wise, this is a big one...  
 First, a-lot-of-optimizations:
@@ -23,7 +23,7 @@ Plus inevitable bug-fixes...
 
 [Home](https://vladmandic.github.io/sdnext/) | [ChangeLog](https://github.com/vladmandic/automatic/blob/master/CHANGELOG.md) | [Docs](https://vladmandic.github.io/sdnext-docs/) | [Discord](https://discord.com/invite/sd-next-federal-batch-inspectors-1101998836328697867) | [Sponsor](https://github.com/sponsors/vladmandic)  
 
-### Details for 2026-09-14
+### Details for 2026-09-15
 
 - **Models**
   - [Anima 2.9B Preview v1](https://huggingface.co/yeoj34760/Anima-2.9B)  
@@ -126,6 +126,7 @@ Plus inevitable bug-fixes...
   - compile: keep model compiled state
   - detailer: handling of stop/skip/pause
   - framepack: correct device assignment, thanks @li-lizhe
+  - installer: better handle git detached head
   - json: handle file locks
   - log: ansi color handling
   - lora: cleanup tags